Fix CodeQL alerts: precision, dead code, Path.Join, LINQ refactors (#14)
Fixes the CodeQL alerts surfaced after the initial scan:

| Severity | Rule | Before | After |
|---|---|---:|---:|
| error | cs/loss-of-precision | 2 | 0 |
| error | cs/unused-collection | 1 | 0 |
| warning | cs/useless-assignment-to-local | 3 | 0 |
| warning | cs/useless-cast-to-self | 2 | 0 |
| note | cs/path-combine | 12 | 0 |
| note | cs/catch-of-all-exceptions | 10 | 9* |
| note | cs/linq/missed-select | 4 | 3* |
| note | cs/linq/missed-where | 4 | 1* |
| note | cs/missed-ternary-operator | 1 | 0 |
| **Total** |  | **39** | **13** |

\* Remaining alerts are intentional (catch-at-boundary in MCP tools /
file IO; foreach loops with side effects or multi-use locals;
`TryGetValue` inside BM25 hot loop). They'll be dismissed with reasons
in the GitHub UI separately.

**Highlights:**
- `CachePrefixOptimizer.EstimateSavings` was actually broken —
`queriesPerDay * prefixTokens` would overflow `int` before the `/
1_000_000m` divide for any moderately-sized estimate (e.g., 1M queries ×
10K tokens = 10^10). Cast the first operand to `decimal` early.
- `Path.Combine` -> `Path.Join` everywhere in `Koshi.Agents`.
`Path.Join` is the modern .NET-recommended API: it always concatenates
with separators and never silently drops earlier args (which `Combine`
does if a later arg looks rooted).
- Three small dead-code drops (`deficitRoles`, `trackerEntries`,
`budget`) and two redundant `(float)` casts gone.

**Verification:**
`dotnet build -c Release`: 0 warnings, 0 errors.
`dotnet test`: 89/89 pass.

---------

+++ src/Koshi.Core/Harness/QualityTracker.cs
@@ -61,11 +61,11 @@ public sealed class QualityTracker
 
     /// <summary>Average budget utilization.</summary>
     public float AvgBudgetUtilization => _entries.Count > 0
-        ? (float)_entries.Average(e => e.BudgetUtilization) : 0;
+        ? _entries.Average(e => e.BudgetUtilization) : 0;
 
     /// <summary>Average cache ratio.</summary>
     public float AvgCacheRatio => _entries.Count > 0
-        ? (float)_entries.Average(e => e.CacheRatio) : 0;
+        ? _entries.Average(e => e.CacheRatio) : 0;
 
     /// <summary>Number of turns that used fallback.</summary>
     public int FallbackCount => _entries.Count(e => e.FallbackLevel != FallbackLevel.None);
